+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Fixed
+- thrift plugin: `types_yarpc.go` file, which contains exception annotations is
+  always generated.
 
 ## [1.47.1] - 2020-09-02
 ### Removed